Раскидал тесты. Созданы абстрактный класс для Product, Smartphone, Grass. Создан миксин с __repr__.
